ai-agent-fundamentals/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
backoff==2.2.1
    # via langfuse
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6
    # via
    #   griffe
    #   tqdm
distro==1.9.0
    # via openai
genai-prices==0.0.35
    # via pydantic-ai-slim
googleapis-common-protos==1.71.0
    # via opentelemetry-exporter-otlp-proto-http
griffe==1.14.0
    # via pydantic-ai-slim
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   genai-prices
    #   langfuse
    #   openai
    #   pydantic-ai-slim
    #   pydantic-graph
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
jiter==0.11.1
    # via openai
langfuse==3.8.1
    # via simple-pydantic-agent (pyproject.toml)
logfire-api==4.14.2
    # via pydantic-graph
openai==2.6.1
    # via
    #   langfuse
    #   pydantic-ai-slim
opentelemetry-api==1.38.0
    # via
    #   langfuse
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic-ai-slim
opentelemetry-exporter-otlp-proto-common==1.38.0
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-http==1.38.0
    # via langfuse
opentelemetry-proto==1.38.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.38.0
    # via
    #   langfuse
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.59b0
    # via opentelemetry-sdk
packaging==25.0
    # via langfuse
protobuf==6.33.0
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
pydantic==2.12.3
    # via
    #   genai-prices
    #   langfuse
    #   openai
    #   pydantic-ai-slim
    #   pydantic-graph
pydantic-ai-slim==1.7.0
    # via simple-pydantic-agent (pyproject.toml)
pydantic-core==2.41.4
    # via pydantic
pydantic-graph==1.7.0
    # via pydantic-ai-slim
python-dotenv==1.2.1
    # via simple-pydantic-agent (pyproject.toml)
requests==2.32.5
    # via
    #   langfuse
    #   opentelemetry-exporter-otlp-proto-http
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   openai
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-ai-slim
    #   pydantic-graph
urllib3==2.5.0
    # via requests
wrapt==1.17.3
    # via langfuse
zipp==3.23.0
    # via importlib-metadata
